optimize core types for smart contract size efficiency
Replace oversized i32/u32 types with minimal alternatives across all
crates. Introduces semantic type aliases (StatValue=i16, ManaValue=u8,
RoundValue=u8, IndexValue=u8, CountValue=u8, etc.) in battle/src/types.rs
so future resizing is a one-line change.

Key changes: stats i32→i16, mana/economy i32→u8, round/lives/wins
i32→u8, CardId u32→u16, UnitId u32→u16, rarity u32→u8, indices
u32→u8, battle limits u32→u8, set_id u32→u16. Estimated ~53% reduction
in per-player on-chain storage (168+ bytes saved per game session).

+++ battle/src/tests/priority.rs
@@ -2,7 +2,7 @@ use super::*;
 use crate::battle::{CombatUnit, UnitId};
 use crate::types::*;
 
-fn trigger_ref(source_id: UnitId, ability_index: u32) -> AbilityTriggerRef {
+fn trigger_ref(source_id: UnitId, ability_index: u8) -> AbilityTriggerRef {
     AbilityTriggerRef {
         source_id,
         ability_index,
